# Fold compound expressions with literal arguments into literals