--- a/flow-typed/visitor.js
+++ b/flow-typed/visitor.js
@@ -1,3 +1,0 @@
-declare type Visitor<T> = {
-    visit: (T) => void
-};