--- a/src/style-spec/function/definitions/assertion.js
+++ b/src/style-spec/function/definitions/assertion.js
@@ -23,12 +23,12 @@ const types = {
 class Assertion implements Expression {
     key: string;
     type: Type;
-    inputs: Array<Expression>;
+    args: Array<Expression>;
 
-    constructor(key: string, type: Type, inputs: Array<Expression>) {
+    constructor(key: string, type: Type, args: Array<Expression>) {
         this.key = key;
         this.type = type;
-        this.inputs = inputs;
+        this.args = args;
     }
 
     static parse(args: Array<mixed>, context: ParsingContext): ?Expression {
@@ -40,33 +40,32 @@ class Assertion implements Expression {
 
         const type = types[name];
 
-        const inputs = [];
+        const parsed = [];
         for (let i = 1; i < args.length; i++) {
             const input = parseExpression(args[i], context.concat(i, ValueType));
             if (!input) return null;
-            inputs.push(input);
+            parsed.push(input);
         }
 
-        return new Assertion(context.key, type, inputs);
+        return new Assertion(context.key, type, parsed);
     }
 
     compile(ctx: CompilationContext) {
         const jsType = JSON.stringify(this.type.kind.toLowerCase());
-        const inputs = [];
-        for (const input of this.inputs) {
-            inputs.push(ctx.addExpression(input.compile(ctx)));
+        const args = [];
+        for (const input of this.args) {
+            args.push(ctx.addExpression(input.compile(ctx)));
         }
-        const inputsVar = ctx.addVariable(`[${inputs.join(',')}]`);
+        const inputsVar = ctx.addVariable(`[${args.join(',')}]`);
         return `$this.asJSType(${jsType}, ${inputsVar})`;
     }
 
     serialize() {
-        return [ this.type.kind.toLowerCase() ].concat(this.inputs.map(i => i.serialize()));
+        return [ this.type.kind.toLowerCase() ].concat(this.args.map(i => i.serialize()));
     }
 
-    accept(visitor: Visitor<Expression>) {
-        visitor.visit(this);
-        this.inputs.forEach(input => input.accept(visitor));
+    eachChild(fn: (Expression) => void) {
+        this.args.forEach(fn);
     }
 }