--- a/src/style-spec/function/definitions/literal.js
+++ b/src/style-spec/function/definitions/literal.js
@@ -72,7 +72,7 @@ class Literal implements Expression {
         }
     }
 
-    accept(visitor: Visitor<Expression>) { visitor.visit(this); }
+    eachChild() {}
 }
 
 module.exports = Literal;