--- a/src/style-spec/function/definitions/var.js
+++ b/src/style-spec/function/definitions/var.js
@@ -35,7 +35,7 @@ class Var implements Expression {
         return [this.name];
     }
 
-    accept(visitor: Visitor<Expression>) { visitor.visit(this); }
+    eachChild() {}
 }