--- a/test/integration/expression-tests/constant-folding/to-color/test.json
+++ b/test/integration/expression-tests/constant-folding/to-color/test.json
@@ -1,0 +1,16 @@
+{
+  "expectExpressionType": null,
+  "expression": ["to-color", "red"],
+  "inputs": [[{}, {}]],
+  "expected": {
+    "compiled": {
+      "result": "success",
+      "isFeatureConstant": true,
+      "isZoomConstant": true,
+      "type": "Color"
+    },
+    "outputs": [
+      [1, 0, 0, 1]
+    ]
+  }
+}