When parsing an expression, if its arguments are all Literals, then evaluate it immediately and yield a Literal instead.

Yields a modest perf improvement in `tile_layout_dds`:
 - before: 51.37629 - 52.20937 ms / iteration
 - after: 47.81951 - 49.86712 ms / iteration

More importantly, this means we get validation errors at compile time rather than evaluation time for the common case of color curves defined like `[curve, [linear], [zoom], 0, "red", 10, "bleu mispelled"]`

benchmark | master 2200752 | expressions-fold-constant 0567a8f
--- | --- | ---
**map-load** | 111 ms  | 119 ms 
**style-load** | 128 ms  | 168 ms 
**buffer** | 1,167 ms  | 1,120 ms 
**tile_layout_dds** | 1,329 ms  | 1,255 ms 
**fps** | 60 fps  | 60 fps 
**frame-duration** | 7.1 ms, 1% > 16ms  | 6.9 ms, 2% > 16ms 
**query-point** | 0.98 ms  | 1.24 ms 
**query-box** | 73.43 ms  | 75.72 ms 
**geojson-setdata-small** | 17 ms  | 0 ms 
**geojson-setdata-large** | 88 ms  | 175 ms 
**filter** | Done  | Done