@anandthakker 2017-09-13
Example generated code for a color curve with `"red"` as the stop output value (excerpt of the `curve/linear-color` integration test output)

## before

```js
function e4() {   return "red" }
var v5 = [e4];
function e6() {   return $this.toColor(v5) }
```

## after

```js
var v4 = (new $this.Color(1, 0, 0, 1));
function e5() {   return v4 }
```