@anandthakker 2017-09-15
Rebased, generalized, and updated benchmarks.