@anandthakker 2017-09-16
@jfirebaugh updated to address your notes/questions